feat: show capStoryChart in final results
ref https://github.com/mlevison/high-performance-teams-game/issues/38
