+++ src/components/CapStoryChart.module.css
@@ -1,0 +1,4 @@
+.chart {
+  height: 300px;
+  margin-top: 1rem;
+}
