+++ src/components/FinalResults.tsx
@@ -1,14 +1,17 @@
 import React from 'react';
+import { AppState } from 'state';
+import CapStoryChart from './CapStoryChart';
 
 type Props = {
-  storiesCompleted: number;
+  state: AppState;
 };
 
 export default function Results(props: Props) {
   return (
     <>
       <h2>Results</h2>
-      <p>Completed {props.storiesCompleted} user stories</p>
+      <p>Completed {props.state.result.storiesCompleted} user stories</p>
+      <CapStoryChart rounds={props.state.pastRounds} />
     </>
   );
 }
